Update to use t_ref and new dtype

+++ test/plume_rise/sofiev_2012_test.jl
@@ -12,8 +12,8 @@ di = DomainInfo(
     starttime, endtime;
     lonrange = deg2rad(-115):deg2rad(1):deg2rad(-68.75),
     latrange = deg2rad(25):deg2rad(1):deg2rad(53.7),
-    levrange = 1:15,
-    dtype = Float64)
+    levrange = 1:15
+)
 
 puff = Puff(di)
